+++ CHANGELOG.md
@@ -18,6 +18,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Removed
 
+- SaveStateToPlayerPrefs method
+- LoadStateFromPlayerPrefs method
+
 ## [3.0.0-beta2] 2025-04-09
 
 ### Added